# use GOMOD && go test -v && adjust bench_test