--- a/tidb/mergesort/Makefile
+++ b/tidb/mergesort/Makefile
@@ -1,10 +1,12 @@
+GO        := GO111MODULE=on go
+
 .PHONY: all
 
 all: test bench
 
 test:
-	go test
+	$(GO) test -v
 
 bench:
-	go test -bench Benchmark -run xx -count 5 -benchmem
+	$(GO) test -bench Benchmark -run xx -count 5 -check.b