@zz-jason 2019-04-19
> golang has not supported about setup suite

Yes, this is not supported in the golang testing. But we can use `b.ResetTimer()` to ignore the time spent on the prepare stage before benchmark test. So I think this is not the correct reason to use `github.com/pingcap/check`. Anyway, I do agree that `github.com/pingcap/check` is much convenient to write benchmarks if we have a lot of benchmark tests.

But in this scenario, we only have two benchmarks. As you can see in this PR, in order to use `github.com/pingcap/check`, the code is more complicated.

> Avoid misleading, I delete the benchmem flag.

Actually, I need the result of memory benchmark. So it's better to have it back.

> close #61

To close issue #61, I think the minimal change is to only replace the benchmark input to a fixed one.