@xiekeyi98 2019-04-19
@zz-jason, Well, thanks for your guidance.
I will think about other ways to make it elegant.